fix(mobile): received user in BottomsTabNavigator bb-622
